Move lots of things around.
