+++ package.json
@@ -36,6 +36,7 @@
     "@nightingale-elements/nightingale-structure": "5.2.0",
     "@nightingale-elements/nightingale-track": "5.2.0",
     "@nightingale-elements/nightingale-variation": "5.2.0",
+    "color-hash": "^2.0.2",
     "core-js": "^3.13.0",
     "data-loader": "^2.9.1",
     "file-saver": "^2.0.2",
