+++ src/adapters/variation-graph-adapter.ts
@@ -1,10 +1,10 @@
-export const transformData = (data) => {
+const transformData = (data) => {
   if (data.sequence && data.features.length) {
     const variants = data.features.map((variant) => ({
-        ...variant,
-        accession: variant.genomicLocation?.join(", "),
-        start: variant.begin,
-      }));
+      ...variant,
+      accession: variant.genomicLocation?.join(', '),
+      start: variant.begin,
+    }));
 
     const total = new Uint8ClampedArray(data.sequence.length);
     const diseaseTotal = new Uint8ClampedArray(data.sequence.length);
@@ -28,20 +28,27 @@ export const transformData = (data) => {
     }
 
     const graphData = [
-        {
-          name: 'variant',
-          range: [0, Math.max(Math.max(...total), Math.max(...diseaseTotal))],
-          color: 'darkgrey',
-          values: [...total].map((value, index) => ({ position: index, value: value })),
-        },
-        {
-          name: 'disease causing variant',
-          range: [0, Math.max(Math.max(...total), Math.max(...diseaseTotal))],
-          color: 'red',
-          values: [...diseaseTotal].map((value, index) => ({ position: index, value: value })),
-        },
-      ];
+      {
+        name: 'variant',
+        range: [0, Math.max(Math.max(...total), Math.max(...diseaseTotal))],
+        color: 'darkgrey',
+        values: [...total].map((value, index) => ({
+          position: index,
+          value: value,
+        })),
+      },
+      {
+        name: 'disease causing variant',
+        range: [0, Math.max(Math.max(...total), Math.max(...diseaseTotal))],
+        color: 'red',
+        values: [...diseaseTotal].map((value, index) => ({
+          position: index,
+          value: value,
+        })),
+      },
+    ];
     return graphData;
   }
 };
 
+export default transformData;
