--- src/types/common-types.ts
@@ -1,25 +1,0 @@
-export type AlphafoldPayload = Array<{
-  entryId: string;
-  gene: string;
-  uniprotAccession: string;
-  uniprotId: string;
-  uniprotDescription: string;
-  taxId: number;
-  organismScientificName: string;
-  uniprotStart: number;
-  uniprotEnd: number;
-  uniprotSequence: string;
-  modelCreatedDate: string;
-  latestVersion: number;
-  allVersions: number[];
-  isReviewed: boolean;
-  isReferenceProteome: boolean;
-  cifUrl?: string;
-  bcifUrl?: string;
-  amAnnotationsUrl?: string;
-  amAnnotationsHg19Url?: string;
-  amAnnotationsHg38Url?: string;
-  pdbUrl: string;
-  paeImageUrl: string;
-  paeDocUrl: string;
-}>;
